Fix lints on Darwin
Fixes lint errors lke these:

    Error: pkg/apis/k0s/v1beta1/cplb.go:143:4: SA4023(related information): the lhs of the comparison is the 2nd return value of this function call (staticcheck)
    			k.VRRPInstances[i].Interface, err = getNIC(k.VRRPInstances[i].Interface)
    			^
    Error: pkg/apis/k0s/v1beta1/cplb.go:144:7: SA4023: this comparison is always true (staticcheck)
    			if err != nil {
    			   ^
    Error: pkg/supervisor/supervisor.go:244:2: SA4023(related information): the lhs of the comparison is the 2nd return value of this function call (staticcheck)
    	ph, err := openPID(p)
    	^
    Error: pkg/supervisor/supervisor.go:245:5: SA4023: this comparison is always true (staticcheck)
    	if err != nil {
    	   ^
