+++ pkg/apis/k0s/v1beta1/cplb_other.go
@@ -6,6 +6,7 @@
 package v1beta1
 
 import (
+	"errors"
 	"fmt"
 	"runtime"
 )
@@ -14,6 +15,6 @@ func getDefaultNIC() (string, error) {
 	return "", fmt.Errorf("getDefaultNIC on %s is not supported", runtime.GOOS)
 }
 
-func getNIC(_ string) (string, error) {
-	return "", fmt.Errorf("getNIC on %s is not supported", runtime.GOOS)
+func macToInterfaceName(_ *string, errs *[]error) {
+	*errs = append(*errs, fmt.Errorf("%w on %s: resolving interface names for MAC addresses", errors.ErrUnsupported, runtime.GOOS))
 }
